Moved handler cache into new module
